Restructure Heroic Platform
