+++ src/platforms/heroic/heroic_game.rs
