+++ src/platforms/heroic/heroic_game_type.rs
