+++ src/platforms/heroic/settings.rs
