+++ src/ui/ui_settings.rs
@@ -2,9 +2,6 @@ use copypasta::ClipboardProvider;
 use eframe::egui;
 use egui::ScrollArea;
 
-#[cfg(target_family = "unix")]
-use crate::heroic::HeroicPlatform;
-
 use super::{
     ui_colors::{BACKGROUND_COLOR, EXTRA_BACKGROUND_COLOR},
     MyEguiApp,
@@ -37,11 +34,6 @@ impl MyEguiApp {
                     platform.render_ui(ui);
                     ui.add_space(SECTION_SPACING);
                 }
-
-                #[cfg(target_family = "unix")]
-                {
-                    self.render_heroic_settings(ui);
-                }
                 self.render_legendary_settings(ui);
                 self.render_lutris_settings(ui);
                 
@@ -54,60 +46,7 @@ impl MyEguiApp {
     }
     
 #[cfg(target_family = "unix")]
-    fn render_heroic_settings(&mut self, ui: &mut egui::Ui) {
-        ui.heading("Heroic");
-        ui.checkbox(&mut self.settings.heroic.enabled, "Import from Heroic");
-        ui.checkbox(&mut self.settings.heroic.default_launch_through_heroic, "Always launch games through Heroic");
-         let safe_mode_header = match (self.settings.heroic.default_launch_through_heroic,self.settings.heroic.launch_games_through_heroic.len()) {
-                        (false,0) => "Force games to launch through Heroic Launcher".to_string(),
-                        (false,1) => "One game forced to launch through Heroic Launcher".to_string(),
-                        (false,x) => format!("{} games forced to launch through Heroic Launcher", x),
-            
-                        (true,0) => "Force games to launch directly".to_string(),
-                        (true,1) => "One game forced to launch directly".to_string(),
-                        (true,x) => format!("{} games forced to launch directly", x),
-            
-                    };
-        egui::CollapsingHeader::new(safe_mode_header)
-                .id_source("Heroic_Launcher_safe_launch")
-                .show(ui, |ui| {
-if 
-self.settings.heroic.default_launch_through_heroic{
-                                   ui.label("Some games work best when launched directly, select those games below and BoilR will create shortcuts that launch the games directly.");
-                    
-                }   else{
-                                   ui.label("Some games must be started from the Heroic Launcher, select those games below and BoilR will create shortcuts that opens the games through the Heroic Launcher.");
-                    
-                }
-                #[cfg(target_family = "unix")]{
-                  let manifests =self.heroic_games.get_or_insert_with(||{
-                        let heroic_setting = self.settings.heroic.clone();
-        
-                        let heroic_platform =HeroicPlatform{
-                        settings:heroic_setting
-                    };
-                        heroic_platform.get_heroic_games()                        
-                    });
-                                    
-                    let safe_open_games = &mut self.settings.heroic.launch_games_through_heroic;
-                    for manifest in manifests{
-                        let key = manifest.app_name();
-                        let display_name = manifest.title();
-                        let mut safe_open = safe_open_games.contains(&display_name.to_string()) || safe_open_games.contains(&key.to_string());
-                        if ui.checkbox(&mut safe_open, display_name).clicked(){
-                            if safe_open{
-                                safe_open_games.push(key.to_string());
-                            }else{
-                                safe_open_games.retain(|m| m!= display_name && m!= key);
-                            }
-                        }                  
-                        }
-                    }
-                })        ;
-        ui.add_space(SECTION_SPACING);
-    }
-
-    fn render_lutris_settings(&mut self, ui: &mut egui::Ui) {
+     fn render_lutris_settings(&mut self, ui: &mut egui::Ui) {
         ui.heading("Lutris");
         ui.checkbox(&mut self.settings.lutris.enabled, "Import from Lutris");
         if self.settings.lutris.enabled {
