+++ src/ui/uiapp.rs
@@ -1,8 +1,5 @@
 use std::{collections::HashMap, error::Error};
 
-#[cfg(target_family = "unix")]
-use crate::heroic::HeroicGameType;
-
 use eframe::{egui, App, Frame};
 use egui::{ImageButton, Rounding, Stroke, TextureHandle};
 use steam_shortcuts_util::shortcut::ShortcutOwned;
@@ -44,8 +41,6 @@ pub struct MyEguiApp {
     ui_images: UiImages,
     pub(crate) games_to_sync: Receiver<FetcStatus<Vec<(String, Vec<ShortcutOwned>)>>>,
     pub(crate) status_reciever: Receiver<SyncProgress>,
-    #[cfg(target_family = "unix")]
-    pub(crate) heroic_games: Option<Vec<HeroicGameType>>,
     pub(crate) image_selected_state: ImageSelectState,
     pub(crate) backup_state: BackupState,
     pub(crate) disconect_state: DiconnectState,
@@ -64,9 +59,7 @@ impl MyEguiApp {
             rt: runtime,
             games_to_sync: watch::channel(FetcStatus::NeedsFetched).1,
             ui_images: UiImages::default(),
-            status_reciever: watch::channel(SyncProgress::NotStarted).1,
-            #[cfg(target_family = "unix")]
-            heroic_games: None,
+            status_reciever: watch::channel(SyncProgress::NotStarted).1,            
             image_selected_state: ImageSelectState::default(),
             backup_state: BackupState::default(),
             disconect_state: DiconnectState::default(),
